Added NULL checks in PHP files
Prevents NULL values form being added from HTML on certain requirements.

+++ delcust-update.php
@@ -17,32 +17,39 @@
 			 $lastname = $_POST['lastname'];
 			 $email  = $_POST['email'];
 			 
+			if($id==NULL){
+			 	echo "Field cannot be NULL. Request failed. Please try again.<br><br>";
+
+			 	echo "<a style='color:black' href='http://localhost:8888/Library-Database/deletecustomer.html'>Back</a>";
+			 } 
+			 
+			 else{
+
+				// Create connection
+				$conn = new mysqli($servername, $username, $password, $dbname);
+
+				// Check connection
+				if ($conn->connect_error) {
+				    die("Connection failed: " . $conn->connect_error);
+				} 
+
+
+				// SQL statement
+				$sql = "DELETE FROM customer 
+						WHERE Customer_ID = '$id'";
+
+
+				echo "Record deleted<br>";
+
+
+				if ($conn->query($sql) !== TRUE) {
+		    		echo "Error: " . $sql . "<br>" . $conn->error;
+				}
 			
+				echo "<a style='color:black' href='http://localhost:8888/Library-Database/'>Home</a>";
 
-			// Create connection
-			$conn = new mysqli($servername, $username, $password, $dbname);
-
-			// Check connection
-			if ($conn->connect_error) {
-			    die("Connection failed: " . $conn->connect_error);
-			} 
-
-
-			// SQL statement
-			$sql = "DELETE FROM customer 
-					WHERE Customer_ID = '$id'";
-
-
-			echo "Record deleted<br>";
-
-
-			if ($conn->query($sql) !== TRUE) {
-	    		echo "Error: " . $sql . "<br>" . $conn->error;
+				$conn->close();			
 			}
-		
-			echo "<a style='color:black' href='http://localhost:8888/Library-Database/'>Home</a>";
-
-			$conn->close();			
 		?>
 		
 	</body>
